+++ src/client/FragPool.cc
@@ -64,7 +64,7 @@ void FragPool::draw( const Frag* frag )
 
     tf.colour.w.w = flags & FADEOUT_BIT ? clamp( frag->life, 0.0f, 1.0f ) : 1.0f;
 
-    model->schedule( -1 );
+    model->schedule( -1, Mesh::SCENE_QUEUE );
 
     tf.colour.w.w = 1.0f;
   }