+++ src/client/ui/Inventory.hh
@@ -39,7 +39,7 @@ class Inventory : public Frame
   private:
 
     static const int   COLS          = 8;
-    static const int   PADDING_SIZE  = 2;
+    static const int   PADDING_SIZE  = 1;
     static const int   FOOTER_SIZE   = 32;
     static const int   ICON_SIZE     = 32;
     static const int   SLOT_SIZE     = 68;